requirement.txt
bcrypt==4.2.0
cffi==1.17.0
cryptography==43.0.0
paramiko==3.4.1
pycparser==2.22
PyNaCl==1.5.0
pytz==2024.1
PyYAML==6.0.2
tqdm==4.66.5
